feat(home): add new sections for core features, core values, CTA, and hero to enhance user engagement and information delivery
- Introduce `CoreFeaturesSection` to showcase key features of the product.
- Add `CoreValuesSection` to communicate the principles guiding the product.
- Implement `CTASection` for encouraging user sign-ups for early access.
- Create `HeroSection` to present a compelling introduction and call to action.
- Refactor `HomePage` to utilize these new components for better organization and readability.
