+++ app/(logged-out)/home/components/cta-section.tsx
@@ -1,0 +1,68 @@
+'use client';
+
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import { motion } from 'framer-motion';
+import { Code2, Sparkles } from 'lucide-react';
+
+interface CTASectionProps {
+  onApplyClick: () => void;
+}
+
+export function CTASection({ onApplyClick }: CTASectionProps) {
+  return (
+    <section id="cta-section" className="py-16 sm:py-32 pb-12 sm:pb-24 px-4 sm:px-6">
+      <div className="container mx-auto text-center max-w-3xl">
+        <motion.div
+          initial={{ opacity: 0, y: 20 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          viewport={{ once: true }}
+          transition={{ duration: 0.6 }}
+        >
+          <div className="mb-6 sm:mb-8">
+            <Badge
+              variant="outline"
+              className="px-2 sm:px-3 py-1 text-xs font-mono bg-emerald-500/10 border-emerald-500/20 text-emerald-600 hover:bg-emerald-500/20 hover:border-emerald-500/40 transition-all duration-300 cursor-default relative overflow-hidden"
+            >
+              {/* Shine effect */}
+              <motion.div
+                className="absolute inset-0 -top-1 -bottom-1 bg-gradient-to-r from-transparent via-white/20 to-transparent"
+                initial={{ x: '-100%' }}
+                animate={{ x: '200%' }}
+                transition={{
+                  duration: 2,
+                  repeat: Infinity,
+                  repeatDelay: 3,
+                  ease: 'easeInOut',
+                }}
+              />
+              <Sparkles className="w-3 h-3 mr-1" />
+              Private Alpha
+            </Badge>
+          </div>
+
+          <h2 className="text-3xl sm:text-4xl lg:text-5xl font-bold mb-4 sm:mb-6 font-mono">
+            Join the Private Alpha
+          </h2>
+
+          <p className="text-base sm:text-lg text-muted-foreground mb-8 sm:mb-12 font-sans px-2">
+            Want to use a tool that fixes what others ignore? We read every reply. Real feedback
+            shapes what we build next.
+          </p>
+
+          <div className="flex flex-col sm:flex-row gap-3 sm:gap-4 justify-center">
+            <Button
+              size="lg"
+              variant="outline"
+              className="w-full sm:w-auto px-6 sm:px-8 py-3 font-mono"
+              onClick={onApplyClick}
+            >
+              <Code2 className="mr-2 h-4 w-4" />
+              Apply for Early Access
+            </Button>
+          </div>
+        </motion.div>
+      </div>
+    </section>
+  );
+}
